# Fix unsafe pointer usage in MurmurBytes