This commit updates MurmurBytes to read the input data as a block of []uint32 rather than one uint32 at a time.  This allows us to signal to the compiler that the data is aligned in a way that doesn't fail the checkptr check.

This commit also swaps in `bits.RotateLeft32` instead of the handrolled version as it is equally fast and there is no need to duplicate stdlib

Profiling revealed that creating the `tail` slice was actually a source of some CPU usage so this commit also removes that and instead calculates the length of the tail and the index where it starts.

I ran the benchmarks 50 times and saved the results as `original.txt` and then I ran the new code 50 times, twice:

```
$ benchstat original.txt new.txt
name               old time/op  new time/op  delta
100MurmurBytes-2   29.6µs ± 4%  29.0µs ± 3%   -2.03%  (p=0.000 n=46+46)
100Murmur64-2       372ns ± 3%   384ns ± 2%   +3.39%  (p=0.000 n=49+50)
100MurmurString-2   924ns ±16%   829ns ± 3%  -10.18%  (p=0.000 n=49+48)
MurmurStringBig-2  33.6ms ± 9%  30.8ms ±10%   -8.27%  (p=0.000 n=45+49)

$ benchstat original.txt new1.txt
name               old time/op  new time/op  delta
100MurmurBytes-2   29.6µs ± 4%  29.7µs ± 9%    ~     (p=0.567 n=46+45)
100Murmur64-2       372ns ± 3%   400ns ± 5%  +7.49%  (p=0.000 n=49+44)
100MurmurString-2   924ns ±16%   923ns ± 5%  -0.07%  (p=0.017 n=49+49)
MurmurStringBig-2  33.6ms ± 9%  33.3ms ± 4%    ~     (p=0.153 n=45+49)
```

The benchmarks vary but these indicate that the new code is at least as fast as the original if not slightly faster