@joepin 2020-03-16
Just to be safe, I re-ran the benchmarks 100 times on each of `master`, `joey/murmur`, and `kevinconaway/murmur-perf` branches, and compared the results:

```
$ benchstat master.txt joey.txt 
name               old time/op  new time/op  delta
100MurmurBytes-8   30.0µs ± 6%  35.0µs ± 7%  +16.87%  (p=0.000 n=99+92)
100Murmur64-8       373ns ± 8%   394ns ± 8%   +5.57%  (p=0.000 n=98+98)
100MurmurString-8   903ns ±12%  1008ns ± 8%  +11.64%  (p=0.000 n=100+97)
MurmurStringBig-8  35.1ms ± 4%  40.0ms ± 4%  +14.03%  (p=0.000 n=99+98)
```

```
$ benchstat master.txt kevin.txt 
name               old time/op  new time/op  delta
100MurmurBytes-8   30.0µs ± 6%  28.6µs ± 6%  -4.57%  (p=0.000 n=99+100)
100Murmur64-8       373ns ± 8%   382ns ±10%  +2.31%  (p=0.001 n=98+100)
100MurmurString-8   903ns ±12%   829ns ± 9%  -8.12%  (p=0.000 n=100+100)
MurmurStringBig-8  35.1ms ± 4%  36.2ms ± 4%  +3.19%  (p=0.000 n=99+99)
```

```
$ benchstat joey.txt kevin.txt 
name               old time/op  new time/op  delta
100MurmurBytes-8   35.0µs ± 7%  28.6µs ± 6%  -18.35%  (p=0.000 n=92+100)
100Murmur64-8       394ns ± 8%   382ns ±10%   -3.09%  (p=0.000 n=98+100)
100MurmurString-8  1.01µs ± 8%  0.83µs ± 9%  -17.70%  (p=0.000 n=97+100)
MurmurStringBig-8  40.0ms ± 4%  36.2ms ± 4%   -9.51%  (p=0.000 n=98+99)
```